requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==4.2.7
django-ckeditor-5==0.2.10
django-countries==7.5.1
django-multipleselectfield==0.2.0.post2
django-quill-editor==0.1.40
django-summernote==0.8.20.0
Pillow==10.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
webencodings==0.5.1
